Add files via upload
This research paper, published in the International Journal of Advanced Science and Technology (2020), explores the growing issue of e-commerce frauds in India, driven by the rapid adoption of digital payments. The paper highlights the challenges businesses face in detecting and preventing fraud, emphasizing the need for advanced technological tools to manage these risks effectively.